逻辑bug修改

+++ config/src/main/java/org/game/config/entity/Item.java
@@ -4,5 +4,5 @@ public interface Item {
 
     int id();
 
-    int num();
+    long num();
 }
