+++ config/src/main/java/org/game/config/entity/SimpleItem.java
@@ -1,4 +1,4 @@
 package org.game.config.entity;
 
-public record SimpleItem(int id, int num) implements Item {
+public record SimpleItem(int id, long num) implements Item {
 }
